+++ packages/connect/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @klever/connect
 
+## 0.1.9
+
+### Patch Changes
+
+- Updated dependencies [bde148e]
+  - @klever/connect-transactions@0.1.8
+  - @klever/connect-contracts@0.2.4
+  - @klever/connect-wallet@0.2.4
+
 ## 0.1.8
 
 ### Patch Changes